# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rabbit Hole",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/agent": {
      "post": {
        "summary": "Respond",
        "operationId": "respond_api_agent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent text SSE v2; see docs/API.md",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Invalid, expired or incompatible continuation"
          },
          "413": {
            "description": "Request body exceeds configured byte limit"
          },
          "429": {
            "description": "Rate or concurrent job limit"
          },
          "503": {
            "description": "Model configuration is missing"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/title": {
      "post": {
        "summary": "Title",
        "operationId": "title_api_title_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TitleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TitleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/structure": {
      "post": {
        "summary": "Structure",
        "operationId": "structure_api_structure_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructureResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "delete": {
        "summary": "Cancel",
        "operationId": "cancel_api_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentRequest": {
        "properties": {
          "attachment_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 4,
            "title": "Attachment Ids"
          },
          "requested_tool": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "web_search",
                  "read_page"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Tool"
          },
          "node_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NodeContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "query": {
            "type": "string",
            "maxLength": 2000,
            "title": "Query",
            "default": ""
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "continuation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000000
              },
              {
                "type": "null"
              }
            ],
            "title": "Continuation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "AgentRequest"
      },
...
      "CardPresentation_TextSpan_": {
        "properties": {
          "heading": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Heading"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroundedText_TextSpan_"
              },
              {
                "type": "null"
              }
            ]
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/CardSection_TextSpan_"
            },
            "type": "array",
            "maxItems": 4,
            "title": "Sections"
          },
          "table": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CardTable_TextSpan_"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "heading",
          "summary",
          "sections",
          "table"
        ],
        "title": "CardPresentation[TextSpan]"
      },
...
      "EntityExtract": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "subtype": {
            "type": "string",
            "enum": [
              "concept",
              "technology",
              "company",
              "product",
              "person",
              "organization",
              "service",
              "software",
              "model",
              "standard",
              "method",
              "field",
              "event",
              "place",
              "country",
              "work",
              "material",
              "species",
              "metric",
              "dataset",
              "policy",
              "project",
              "language",
              "other"
            ],
            "title": "Subtype"
          },
          "qualifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualifier"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "role": {
            "type": "string",
            "enum": [
              "main",
              "related"
            ],
            "title": "Role"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/TextSpan"
            },
            "type": "array",
            "maxItems": 4,
            "title": "References"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/EntityLink"
            },
            "type": "array",
            "title": "Links"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "name",
          "subtype",
          "qualifier",
          "aliases",
          "role",
          "links"
        ],
        "title": "EntityExtract"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InformationExtract": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "subtype": {
            "type": "string",
            "enum": [
              "concept",
              "entity",
              "claim",
              "example",
              "comparison",
              "procedure"
            ],
            "title": "Subtype"
          },
          "title": {
            "$ref": "#/components/schemas/TextSpan"
          },
          "excerpt": {
            "$ref": "#/components/schemas/TextSpan"
          },
          "presentation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CardPresentation_TextSpan_"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "subtype",
          "title",
          "excerpt"
        ],
        "title": "InformationExtract"
      },
...
      "NodeContext": {
        "properties": {
          "node_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Node Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "information",
              "source",
              "entity"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "maxLength": 500,
            "title": "Title"
          },
          "text": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node_id",
          "kind",
          "title",
          "text"
        ],
        "title": "NodeContext"
      },
...
      "StructureRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "continuation": {
            "type": "string",
            "maxLength": 2000000,
            "minLength": 1,
            "title": "Continuation"
          },
          "text_hash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Text Hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "continuation",
          "text_hash"
        ],
        "title": "StructureRequest"
      },
      "StructureResult": {
        "properties": {
          "version": {
            "type": "integer",
            "enum": [
              1,
              2,
              3
            ],
            "title": "Version",
            "default": 1
          },
          "text_hash": {
            "type": "string",
            "title": "Text Hash"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InformationExtract"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Items"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/EntityExtract"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Entities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text_hash",
          "items"
        ],
        "title": "StructureResult"
      },
      "TextSpan": {
        "properties": {
          "start": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start"
          },
          "end": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "End"
          },
          "quote": {
            "type": "string",
            "title": "Quote"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start",
          "end",
          "quote"
        ],
        "title": "TextSpan"
      },
      "TitleRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "continuation": {
            "type": "string",
            "maxLength": 2000000,
            "minLength": 1,
            "title": "Continuation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "continuation"
        ],
        "title": "TitleRequest"
      },
      "TitleResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TitleResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}